Make connection establishing smarter
There were a couple cases where we need to be optimistic about
preserving existing connection setups when `#establish_connection`
is called:

1. Some gems are paranoid and call `#remove_connection` and
   `#establish_connection` (`activerecord-import` is the primary
   offender for this).
2. Rails will remove and re-restablish connections in Rake tasks.

For both of these cases we want to be optimistic and preserve our
existing primary-replica configuration if the new connection
specification matches the primary pool's specification. This is so
that our primary-replica configuration doesn't get tossed out every
time a spurious remove-and-establish happens.